Introduce a defined sort order
Order is according to an order index number. Tasks that are created later
should be displayed later in the list. Thus, the current millisecond is
used as an order index. A random value is added in order to have a defined
order between Tasks that are created at the same millisecond.